+++ graphics/vulkan/r_vulkan.h
@@ -9,6 +9,7 @@
 #include "r_dtypes.h"
 #include <SDL2/SDL_vulkan.h>
 
+#define HAS_COMPUTE
 #define MAX_WRITE_DESCRIPTORS 32
 
 namespace Fresa::Graphics